[fix] Surface RBAC roles in the OSS frontend
RBAC enforcement is on in OSS, but roles were invisible in the OSS UI:

- Move the GET /access/roles endpoint from the EE access router to OSS
  (it uses get_roles/SCOPES, which now live in OSS). EE keeps /plans and
  inherits /roles via its additive mount, so both editions serve it.
- Web members UI gated role display/selection behind `isEE() && hasRBAC`,
  which is false in OSS. Switch the gate to `!isEE() || hasRBAC` so OSS
  always shows the role column, the populated invite-role selector, and
  the edit-role dropdown; EE keeps the plan gate (not-entitled EE still
  hides roles and shows the upsell).
